perf: Render program more efficiently
Previously, the entire program was rendered for every frame. For large
programs, this caused the TUI to waste a lot of time rendering what
would ultimately be discarded, due to most of the program being
off-screen, severely impacting usability.

Now, the relevant information about the program is pre-computed and
cached. From this cache, only the relevant lines are fetched and
rendered.